Simplify the `wrapRegExp` helper for named groups (#13201)
